fixed a float16 oob bug in internode test

+++ exps/grpcoll/test_internode_grpcoll.py
@@ -99,12 +99,17 @@ def prepare_test_func_kwargs(
     head_dim = hidden_size // num_heads
     x = torch.ones((num_tokens, hidden_size), dtype=dtype, device="cuda")
     if distinct_token:
-        x *= torch.arange(
-            rank * num_tokens,
-            (rank + 1) * num_tokens,
-            dtype=dtype,
-            device="cuda",
-        ).view(-1, 1) / (num_ranks * num_tokens)
+        x *= (
+            torch.arange(
+                rank * num_tokens,
+                (rank + 1) * num_tokens,
+                dtype=torch.int64,
+                device="cuda",
+            )
+            .view(-1, 1)
+            .double()
+            / (num_ranks * num_tokens)
+        ).to(dtype)
     else:
         x *= rank
     if cast_lse:
@@ -1138,7 +1143,7 @@ def test_main(
     assert 1 <= num_data_groups <= 3
 
     # choose dtype from {torch.bfloat16, torch.float16, torch.float32, torch.float64}
-    dtype = torch.bfloat16  # TODO: support other dtypes
+    dtype = torch.bfloat16
     assert dtype in (torch.bfloat16, torch.float16, torch.float32, torch.float64)
 
     # Remake the hidden size to control
